mcp_web_interface/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dj-database-url==3.0.0
Django==5.2.3
flatbuffers==25.2.10
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jh2==5.0.9
jiter==0.10.0
markdown2==2.5.3
niquests==3.14.1
numpy==2.3.1
openai==1.93.0
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.1
packaging==25.0
pandas==2.3.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
qh3==1.5.3
requests==2.32.4
requests-cache==1.2.1
retry-requests==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.13.900
uvicorn==0.35.0
wassima==1.2.2